requirements.txt
annotated-types==0.5.0
blinker==1.6.2
click==8.1.4
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
Flask-PyMongo==2.3.0
flask-sqlacodegen==2.0.0
greenlet==2.0.2
inflect==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.0.2
pydantic_core==2.1.2
pymongo==4.4.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
python-dotenv